feat(contextmenu): remove need to inject service to app ctrl
Before this commit an app controller was needed to connect a contextmenu
with the contextmenu service via a unique variable. The controller code
looked something like this:

```js
AppCtrl.$inject = ['ContextmenuService'];
function AppCtrl($contextmenu) {
  var vm = this;
  vm.contextmenu = $contextmenu;
}
```
The HTML referenced this variable via the contextmenu-paramter:

```html
<div ng-controller="AppCtrl as appCtrl">
  <div contextmenu="appCtrl.contextmenu">
  ...
  </div>
</div>
```

With this commit the need to use a controller to link the menu to a service
is removed. The HTML code with a reference to a controller variable is still
needed. This makes it possible to use multiple contextmenu's on the same
page on different elements.